// apps/dashboard/src/types/exceptionMessage.ts
export type ERROR_CODES =
  | 'USER_NOT_FOUND'
  | 'PASSWORD_INVALID'
  | 'PRODUCT_NOT_FOUND'
  | 'PRODUCT_DISCOUNT_PRICE_IS_BIGGER_THAN_PRICE'
  | 'PRODUCT_DISCOUNT_PRICE_IS_ZERO'
  | 'PRODUCT_OUT_OF_STOCK'
  | 'PRODUCT_IS_IN_AUTOMATION'
  | 'ORDER_NOT_FOUND'
  | 'ORDER_EXPIRED'
  | 'ORDER_SHOP_IS_NOT_VALID'
  | 'ORDER_INVALID'
  | 'ORDER_QUANTITY_IS_ZERO'
  | 'ORDER_CARD_TO_CARD_NOT_UPLOADED'
  | 'ORDER_PENDING_EXIST'
  | 'ORDER_STATUS_NOT_PENDING'
  | 'ORDER_STATUS_NOT_PAYMENT'
  | 'ORDER_SHOULD_HAVE_PRODUCTS'
  | 'ORDER_SHOULD_HAVE_CARD_TO_CARD'
  | 'ORDER_HAS_ALREADY_BEEN_SUBMITTED'
  | 'LEAD_INSTAGRAM_NOT_FOUND'
  | 'NO_INSTAGRAM'
  | 'INSTAGRAM_REQUIRED'
  | 'CITY_INVALID'
  | 'CITY_NOT_FOUND'
  | 'COMMENT_NOT_FOUND'
  | 'CONTENT_TEXT_REQUIRED'
  | 'CONTENT_FILE_REQUIRED'
  | 'CONTENT_INSTAGRAMPOST_REQUIRED'
  | 'CONTENT_AUDIO_REQUIRED'
  | 'CONTENT_VIDEO_REQUIRED'
  | 'CONTENT_IMAGE_REQUIRED'
  | 'CONTENT_SHOULD_HAVE_AT_LEAST_ONE_PRODUCT'
  | 'FILE_NOT_FOUND'
  | 'CONDITION_IS_ALREADY_EXIST'
  | 'CONSENT_TEXT_REQUIRED'
  | 'USER_DATA_TEXT_REQUIRED'
  | 'REMINDER_TIME_REQUIRED'
  | 'CARD_TO_CARD_USER_HAVE_NOT'
  | 'ORDER_PENDING_NOT_FOUND'
  | 'SHOP_NOT_FOUND'
  | 'REMINDER_MESSAGE_REQUIRED'
  | 'PAYMENT_METHOD_NOT_AVAILABLE'
  | 'ZARINPAL_ERROR'
  | 'ZARINPAL_USER_HAVE_NOT'
  | 'ZARINPAL_VERIFY_ERROR'
  | 'VERIFY_TRANSACTION_NOT_FOUND'
  | 'VERIFY_PAYMENT_CANCELED'
  | 'PAYMENT_METHODS_NOT_FOUND'
  | 'ORDER_QUANTITY_IS_MAX'
  | 'VERIFY_PAYMENT_AMOUNT_NOT_MATCH'
  | 'PLAN_ALREADY_SUBSCRIBED'
  | 'SUBSCRIPTION_ALREADY_PAID'
  | 'REFERRALCODE_NOT_FOUND'
  | 'PERMISSION_DENIED';

export interface ExceptionMessage {
  message: string;
  statusCode: 400 | 401 | 404 | 500;
  code: ERROR_CODES;
  error: 'Not Found' | 'Bad Request' | 'Internal Server Error' | (string & {});
  data: any;
}